ref: Move Span to Scope
